Inheriting class
Write a C++ program that manages basic and departmental information of employees.

Implement classes to handle basic information

such as name, gender, and age, and departmental information such as department name and designation.

Create a class Employee that inherits from both basic information and department information classes.

Prompt the user to input information for an employee and display the collected information.